# Speed up your Vagrant with this one weird trick!