--- a/Vagrantfile
+++ b/Vagrantfile
@@ -14,6 +14,10 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
   config.vm.synced_folder ".", "/vagrant", type: :nfs
 
   config.vm.provider "virtualbox" do |v|
+
+    v.customize ["modifyvm", :id, "--natdnshostresolver1", "on"]
+    v.customize ["modifyvm", :id, "--natdnsproxy1", "on"]
+
     host = RbConfig::CONFIG['host_os']
 
     # Give VM 1/4 system memory & access to all cpu cores on the host