This speeds up the provisioning of a Vagrant box by enabling a couple of
DNS-related options, and moving the bundle install location outside of
the NFS-mounted canvas root and into ~/canvas_gems.

An ansible plugin has been added that outputs timing information has been
added.

```
PLAY RECAP ********************************************************************
canvas | bundle install ----------------------------------------------- 257.93s
canvas | compile assets ----------------------------------------------- 148.07s
dependencies | Add nodesource ------------------------------------------ 21.64s
canvas | test db setup ------------------------------------------------- 14.38s
update apt cache ------------------------------------------------------- 13.26s
dependencies | Update apt cache ----------------------------------------- 8.88s
dependencies | Update npm to latest ------------------------------------- 8.72s
canvas | wait for all canvas gem files to close ------------------------- 6.23s
canvas | stop jobs ------------------------------------------------------ 6.17s
canvas | npm install ---------------------------------------------------- 5.57s
```
The test database is created and initialized.

In my testing, these changes, along with changing the base box config to have all the OS-level dependencies pre-installed, has reduced `vagrant up` time to ~10 minutes.