@grahamb 2015-08-27
Note: the timing plugin is a python script. It appears that when it gets run, it leaves behind a compiled .pyc file. You may want to add `*.pyc` to either `.git/info/excludes` or your global `.gitignore` file.